Move taxonomy/term related methods to a trait
